voguev210: add rtc work-around
The RTC has issues and may lose time information on power off. That causes problems for login because we use https and when the machine resets, the clock may be set to 1970, so the certificates are not yet valid.
As a workaround, save time when halting, and on boot, if RTC is invalid, use that saved time (default to 1/1/2010 on first boot).

ntp wil set the proper time after a few minutes.

Review URL: http://codereview.chromium.org/1665001