overlay-x86-dogfood: remove agz USE flag.
The dogfood overlay originally got the agz USE flag under the hopes that someday the two overlays would be unified and we'd ship the same build for both platforms. This doesn't seem like it's going to happen, and this hack is impeding someone else's work on agz-specific stuff, so it's going.

BUG=chrome-os-partner:1158
TEST=Adhoc
Build a clean dogfood tree. Hope nothing breaks.

This will not be merged until a long test run is complete.

Change-Id: I6ef63d532b6f0eb622b72ecabb25d53d91bf7d59
Signed-Off-By: Elly Jones <ellyjones@chromium.org>

Review URL: http://codereview.chromium.org/3398020