Script now allows burning a locally-built U-Boot
Looks for u-boot.bin in the current directory and flashes this instead
if it exists. This should be safe because this file would only exist
if the dev had built it.

Also made the script more informative on success/fail.

This makes it quick and easy to try out a new U-Boot

Change-Id: I9b426e15d65a2bf60ab34d96c7023e9c8c0372e6

BUG=chromiumos-9824
TEST=(tried manually it in all 4 combinations)

Review URL: http://codereview.chromium.org/5221002