Make tegra-drivers depend on the correct X server version
The driver binaries in the current BDK package require an X server
with ABI version 5, so these packages will not work on X servers
outside of the 1.6 release branch.

Add RDEPEND entries to enforce that dependency.  Later, when a BDK
package is released with multiple driver ABIs, RDEPEND can be used to
make sure that the appropriate version of the driver is installed
depending on the server version.

Change-Id: Ia693dbf4e062dda4a3bf84c7cc256311bc695587

BUG=none
TEST=Verified that tegra-drivers holds xorg-server at 1.6.3.

Review URL: http://codereview.chromium.org/5974002
Patch from Aaron Plattner <aaronp24@gmail.com>.