Revert "Modify make.conf in board overlays to preserve existing EXTRA_BUILD_ARGS settings"
This reverts commit 4d48f41b361a9c49f579b652c492bba3e9e21c07.

This change does not do what is intended and was improperly tested.

Setting a variable thats in make.conf will prevent the value in make.conf
from ever being seen.  make.conf is not parsed as an actual shell script,
but its parsed by portage.  Apart from special variables like USE, if a
variable is already set portage will not import the value in the config
files and let it be overridden elsewhere.

My testing in the previous CL validated that the new changes showed up
but did not validate that the values in make.conf existed.  I should
have let the build run to copmletion rather than just visually validating
the inputs to the build.

BUG=11809
TEST=Ran setup_board --force and build_packages.

Review URL: http://codereview.chromium.org/6454009

Change-Id: Icb7f47def2cd5c86f6311cf41e1455dd7c31b44d