Select and install appropriate flash device configuration files.
For booting from NAND we have to rely on the existing BCT files
that are checked in to have the correct NAND timings.  This is
a temporary situation and will be resolved once cbootimage can
parse and setup NAND parameters.

BUG=chromium-os:11981
TEST=Build and boot seaboard and harmony using write_tegra_bios.
This ensures that the existing tools are not broken.

Change-Id: Ib474287a959d14f5b91c6088bc1c68ec6df09c74

Review URL: http://codereview.chromium.org/6578027