Add evdev to INPUT_DEVICES for seaboard
There will be a small boottime impact for seaboards that do not use
evdev as they load the driver but fail to find the device to attach to.
As this is not a production release, the tradeoff is against increased
overhead/complexity from forking the seaboard build for one feature with
a relatively minor impact.

Change-Id: I23b4ffa2d35e32d145bd94cb2af32bf761b301f1

BUG=chromium-os:11929
TEST=Ran build_packages, validated evdev added and existing input devices remained
  Unmerged xf86_input_keyboard and xf86_input_mouse, ensured remerged from build_packages
  Unmerged xf86_input_keyboard, xf86_input_mouse, xf86_input_evdev and xorg_drivers, ensured all 4 remreged
  Validated X/UI running on a system which did not use evdev
  Validated X running on a system which did use it

Review URL: http://codereview.chromium.org/6660004