Add runtime dependency to chromeos-flasher-script for all tegra-bct based ebuilds.
This ensures that when the BCT is updated the associated flasher script will
also be updated.

BUG=chromium-os:11981
TEST=emerge tegra-bct for Seaboard and verify that the correct flasher
script is installed and that U-Boot can be written using the existing
cros_write_firmware as well as a new one that reads the flasher script
from it's new location.

Change-Id: Ia1c888ac0266cf72109cc32cbe2f8a716dee2d33

This CL depends on: http://codereview.chromium.org/6730035/

Review URL: http://codereview.chromium.org/6732036