Mario: Disable synaptics driver
We suspect the Synaptics driver may be crashing and causing Smoke suite
failures. Disabling the synaptics driver temporarily can help us figure
out whether it is the cause of the crash.

This is a temporary measure to hopefully open the tree while the crash is
investigated.

Bug to re-enable Synaptics driver: http://crosbug.com/13628

TEST=test on build bot
BUG=chromium-os:13628

R=adlr@chromium.org

Review URL: http://codereview.chromium.org/6773016

Change-Id: I6e273f13be348865f038ff5b93f4425db00e4502