When USE=touchui use rule-based xorg.conf
Note: This is a temporary measure to fork xorg.conf while
performance testing is ongoing for enabling this for all
configurations at which point this will be merged back
into a single file.

Currently bootperf does not work on ARM which is preventing the
validation of this.  As turning on AutoEnableDevices and
AutoAddDevices may have boot time impacts, the timings
need to be validated before deploying for all platforms.

This is going in now to enable testing for other developers
working on the platform.  Without this change the X configuration
is not viable for testing other components with USE=touchui

BUG=NONE
TEST=Emerged with and without USE=touchui, confirmed correct xorg
  Emerged with USE=touchui, confirmed xorg properly fetches
  touchscreen via udev and X autoconfiguration

Review URL: http://codereview.chromium.org/6777030

Change-Id: Ieba70d375f70c4ff718003481002033b107ec120