Chromeos firmware ebuild for tegra2-based boards.
This overrides the chromeos-firmware ebuild from the Chromium OS overlay with
a "9998" numbered version which does nothing to prevent a file collision with
the actual chromeos-firmware-tegra2 ebuild. The "9998" version will be removed
after it is taken out of the depedency graph.

Depends-On: http://gerrit.chromium.org/gerrit/#change,508
BUG=None
TEST=Emerge chromeos-firmware-tegra2 and compare
/usr/sbin/chromeos-firmwareupdate to the version generated from the firmware
ebuild in the tegra2-private overlay that this version is replacing.

Change-Id: I8edfc62229265b63e4a9c1d13ff694ac69bb1bec
Reviewed-on: http://gerrit.chromium.org/gerrit/509
Reviewed-by: Micah Catlin <micahc@google.com>
Reviewed-by: Micah Catlin <micahc@chromium.org>
Reviewed-by: Nick Sanders <nsanders@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>