Add the CHROMEOS_U_BOOT_CONFIG variable to make.conf.
This variable is required by the u-boot-next ebuild. It's first 3 components
are used to generate target names for various types of u-boot images. As far
as I can tell the last 2 components aren't used, but there must be three
underscores for the ebuild to process the setting correctly. The two final
components were chosen to be consistent with the existing uses in ARM.

BUG=chrome-os-partner:3895
TEST=Ran cros_workon_make on u-boot-next for x86-mario, saw it pick up this
setting and attempt to build.

Change-Id: I938cb9eb31dfac1b60c426b130c5d0c62fb90e1a
Reviewed-on: http://gerrit.chromium.org/gerrit/1159
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>