tegra2: make mxt_touch an InputClass in its own .conf snippet
Move mxt_touch touchscreen X input configuration to its own touchscreen.conf snippet.
This file is installed when the touchui USE flag is in use.

The touchscreen has an evdev-compatible kernel driver, and has a /dev/input/event* node.
Thus, an InputClass can be used, to find the correct device, and load the evdev X input driver.

A side affect of this, and the use of AutoAddDevices is the special xorg.conf.touchui becomes redundant and is therefore removed.

BUG=chromium-os:11929
TEST=Atmel maXTouch touchscreen still works for touchui builds.

Change-Id: I582c6ccbb5272f257498c0d0eb23ac3147a8f5a8
Reviewed-on: http://gerrit.chromium.org/gerrit/1110
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>