xorg-conf: Simplify Tegra 2 X.Org configuration
Recent versions of the Tegra X driver auto-detect a lot more things
than it used to.  Many of the options specified in the Tegra-specific
xorg.conf file are no longer used.

Replace the custom xorg.conf files with a minimal one that just
specifies the custom HDMI modes, since libnvdc does not yet provide a
mode list to the X driver.  Add a copy of this section to account for
the fact that it was renamed from "HDMI" to "HDMI-1" in a future
driver releases.  This extra section can be removed once the driver is
updated.

Install xorg.conf.example from the Tegra BSP as
/etc/X11/xorg.conf.d/tegra.conf.  This makes the server load
tegra_drv.so and disables a few extensions that are not useful, such
as DRI and GLX.

BUG=chrome-os-partner:4039
TEST=Verify that the Tegra driver is still loaded and that HDMI hotplug works

Change-Id: I07addffb2498cbdfc713773852d35cad97b30790
Reviewed-on: http://gerrit.chromium.org/gerrit/1344
Tested-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>