Mask tegra-drivers-10, as they require xorg-server-1.10, which isn't available.
Portage always tries to install the latest version of each package that is
stable first. In this case, the latest version of tegra-drivers-10
doesn't work because it requires xorg-server-1.10, which isn't marked
as stable.

Portage 2.1.10 and later actually verify that this is the case and fail the
build if your unmasking is inconsistent. So it's good to fix this.

BUG=chromium-os:16529
TEST=Verify tegra-drivers can be installed with Portage 2.1.10.3

Change-Id: I7280f1c0bf0324f75662947ab3f0960b7b1325a9
Reviewed-on: http://gerrit.chromium.org/gerrit/2637
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: David James <davidjames@chromium.org>