Set an environment variable to select a device tree to put in coreboot's CBFS.
This change adds and configures an environment variable in make.conf to select
a device tree to embed in the coreboot image in its CBFS. The precedence set by
ARM is to build all the device trees all the time, and then select the
appropriate one based either on an explicit selection or board name to actually
embed in the image. This mechanism is different in that it ignores board name,
but it still allows selecting from a collection of device trees.

BUG=chrome-os-partner:4993
TEST=Built coreboot with it's ebuild change and verified that the dtb ended up
in CBFS

Signed-off-by: Gabe Black <gabeblack@google.com>
Change-Id: If8a28c07d28080269d8e42201a3ea0dcf558bd28
Reviewed-on: http://gerrit.chromium.org/gerrit/4708
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Tested-by: Gabe Black <gabeblack@chromium.org>