Hack tegra2_dev-board to think NAND block size is 64K.
Here's the issue we see:

When we build the firmware image (in cros_bundle_firmware), we
end up calling cbootimage to create the 'boot-stub' section.
The cbootimage tool ends up creating a signed binary block
containing:
- Copy 1 of the BCT
- Copy 2 of the BCT
- The R0 copy of u-boot.
One big problem here is that each of the objects above needs
to be aligned to start on a block.  When the block size is 128K,
that means we block out the size we have allocated for
the 'boot-stub' section (~768K)

The right solution is one of the following:
1. Simply don't support NAND for now.  I'm not sure we do such
   a hot job at it right now anyway.  That probably means that
   we should switch our standard build to be seaboard.
2. Add a "NAND" flashmap for u-boot that could potentially give
   more space for boot-stub.
3. Reduce the number of copies of the BCT stored on NAND.
   It's unclear whether both copies are still useful with NAND.

BUG=chromium-os:17464
TEST=compiled

Change-Id: I68cad25c938b1d5a4ad23bd3ffb7dc4d51f929c6
Reviewed-on: http://gerrit.chromium.org/gerrit/4780
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>