New drop of tegra2 ldk for chrome/chromeos-0.0.20
* X11: Tegra X11 driver was updated to perform mode
validation with the kernel, to only expose modes that
can be set. This may reduce the available modes from
the X server's modepool, but all of the remaining
modes should work as expected, instead of failing to
set some modes.

* X11: A bug was fixed that was causing corruption of
the EDID and consequently nondeterministic X mode pools.

* EGL: A bug was fixed that could cause crashes in
libEGL if the eglWaitNative()and eglWaitClient()
functions were called with a pbuffer surface made
current.

* X11: Sanitize foreign pixmap handling.This adds
a special case when destroying foreign pixmaps that
we back with an HW surface such as the screen or
the shadow pixmap. This fixes issues where we would
try to free their surface twice or get them out of
the pixmap linked list when they're not on it. This
also allows us to use real pixmaps to manage the
shadow surface instead of scratch pixmap headers.

* Softfp version is uploaded to crosbug 5377:
http://code.google.com/p/chrome-os-partner/issues/detail?id=5377

* Hardfp version is uploaded to crosbug 5379:
http://code.google.com/p/chrome-os-partner/issues/detail?id=5379

BUG=chrome-os-partner:5377 and chrome-os-partner:5379
TEST=test report attached to partner issue

Change-Id: Id1cda84281310fcb9803ad2ae990392a9f09d5d5
Signed-off-by: Matt Pedro <mapedro@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5516
Reviewed-by: Micah Catlin <micahc@chromium.org>
Tested-by: Micah Catlin <micahc@chromium.org>