xorg-conf: Delete /etc/X11/xorg.conf
The Tegra display driver was previously unable to actually generate a
modepool due to limitations in the kernel interface, so it was
necessary to provide a mode list via xorg.conf.  Those limitations
have since been fixed and the X driver's modepool generation is much
better.  However, these hard-coded modes are still lingering and can
cause problems for some monitors.  Since they're not needed anymore,
and the "DontZap" option doesn't actually work on the version of the X
server we're using, simply delete xorg.conf.

BUG=chrome-os-partner:5441
TEST=Make sure X still starts and HDMI hotplug chooses a reasonable mode

Change-Id: I048657338bf95abeeda9339d5d6d6b42c8e8ebf9
Reviewed-on: http://gerrit.chromium.org/gerrit/6386
Tested-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Robert Morell <rmorell@nvidia.com>