amd64-generic overlay: use a dedicated toolchain tuple
Up until now, we've been re-using the host toolchain in the cros sdk
chroot to build up amd64 boards.  This works for the most part, but
ultimately falls down due to how we inject our custom settings -- via
the sysroot wrapper scripts.  Since we don't wrap the host toolchain,
many of our settings are lost, and many packages fail to build.

The answer is not to add the sysroot wrapper to the host toolchain as
this introduces its own set of problems.  When we go to execute the
host toolchain for compile-time only helpers, our wrapper settings
will also get applied.  Further, while we want to use the gold linker
for the target, we don't want to use it on the host (since there are
packages known to not work with it).

By splitting the amd64 boards off into their own dedicated tuple,
this also allows us to keep the versions independent.  We can update
the target toolchain to a newer version while staying conservative
in the chroot.  Ultimately we want to keep all our versions in sync,
but it's simpler to have them split so we can experiment freely.

BUG=chromium-os:20636
TEST=`./setup_board --board=amd64-generic` produces a x86_64-cros-linux-gnu toolchain
TEST=`./build_packages --board=amd64-generic` now works

Change-Id: I8a733a1909fcc6279909c64eeeafad9c42269491
Reviewed-on: http://gerrit.chromium.org/gerrit/8241
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: asharif <asharif@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>