Be explicit about what firmware implementation is used on vanilla tegra2.
chromeos-firmware-tegra2 depends on a board-specific bct, which should be
defined in all variants of the tegra2 board. It is not defined in vanilla
tegra2, so we cannot use chromeos-firmware-tegra2, and must use the null
implementation instead.

emerge already chooses chromeos-firmware-null on tegra2 -- this just makes
the current behavior explicit so that emerge can do this without
backtracking.

BUG=chromium-os:20981
TEST=Verify build_packages works without backtracking now on tegra2. Confirm
     chromeos-firmware-tegra2 is still used on tegra2 variants.
Change-Id: I438d8b498eaa3c64fec2c6b1bff4c21619678e81
Reviewed-on: http://gerrit.chromium.org/gerrit/8749
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: David James <davidjames@chromium.org>