Rename private opengles to opengles-bin to avoid backtracking.
We want to disable backtracking in emerge so that emerge will always
select the latest version of all packages.

To get this working, we need to ensure that the latest version of all
packages actually builds. On tegra2, when the tegra2-ldk flag is
disabled, we currently rely on backtracking for correct builds. To
prevent this, I've renamed the bin version of opengles to opengles-bin so as
to allow us to choose between them without backtracking.

BUG=chromium-os:20981
TEST=Verify correct merge with --backtrack=0 on tegra2 with tegra2-ldk
     disabled. Also verify right package is merged with tegra2-ldk is
     enabled.

Change-Id: Id94a688ea61529f2ca2b22e4530ef47775b5a10e
Reviewed-on: http://gerrit.chromium.org/gerrit/8796
Reviewed-by: Anton Staaf <robotboy@chromium.org>
Tested-by: David James <davidjames@chromium.org>