Point Waluigi kernel at nVidia git server.
The Waluigi kernel isn't yet in a clean enough state to live on
the Chromium OS git server, so we're just going to point at
the nVidia git servers.  Note that this is not suitable for
shipping Waluigi but is OK for current development.

You can get a cros_workon type functionality if you follow
the instructions in the "kernel-0.0.2.ebuild"

BUG=chromium-os:25131
TEST=emerge-waluigi sys-kernel/waluigi-kernel worked.

Change-Id: Id133513ca1afe4729271019e1405e9831e8f79ce
Reviewed-on: https://gerrit.chromium.org/gerrit/14295
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>