Daisy: Add explicit RDEPEND for exynos-kernel
Since we use a private kernel (sys-kernel/exynos-kernel)
from a public overlay, we need to explicitly RDEPEND
on it as the virtual is not sufficient. For example,
chromeos-kernel (which we do not want to build) could
be used if exynos-kernel is not cros-worked-on as there
is only the exynos-kernel-9999.ebuild (no 0.0.1-r#).

BUG=None
TEST=equery-daisy g chromeos-bsp, verify exynos-kernel
is listed in graph.

Change-Id: Ib74ecd3b97fdef773192666ba28ed132880a7719
Reviewed-on: https://gerrit.chromium.org/gerrit/16867
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Katie Roberts-Hoffman <katierh@chromium.org>
Tested-by: Katie Roberts-Hoffman <katierh@chromium.org>