Daisy: Clean up mfc-fw BCS ebuild
This CL cleans up the mfc-fw ebuild to bring it in line with other BCS
ebuild files. This also adds a comment to show the contents of the BCS tar
file, to provide a better understanding of what this does, and for easier
searching.

BUG=chrome-os-partner:8196
TEST=Manual, verify emerge-daisy mfc-fw completes successfully.

Change-Id: Id46aa5c2030b7e951afcc96e7e6f6072ff1e6485
Reviewed-on: https://gerrit.chromium.org/gerrit/18174
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>