Daisy: Clean up openmax BCS ebuild
This CL cleans up the Daisy openmax ebuild to bring it in line with other BCS
ebuild files. This also adds a comment to show the contents of the BCS tar
file, to provide a better understanding of what this does, and for easier
searching.

BUG=chrome-os-partner:8196
TEST=Manual, verify emerge-daisy openmax completes successfully.

Change-Id: I18f80e1c5312d3b4d8c8b3851e5fb4e5cd0406d1
Reviewed-on: https://gerrit.chromium.org/gerrit/18271
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>