Daisy: Clean up exynos-pre-boot BCS ebuild
This CL cleans up the Daisy exynos-pre-boot ebuild to bring it in line with
other BCS ebuild files. This also adds a comment to show the contents of the
BCS tar file, to provide a better understanding of what this does, and for
easier searching.

BUG=chrome-os-partner:8196
TEST=Manual, verify emerge-daisy exynos-pre-boot completes successfully.

Change-Id: I92221b6811e81ead65ff8cd3bb532dfc5ba77406
Reviewed-on: https://gerrit.chromium.org/gerrit/18270
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>