Set the EAPI version back to 2 for these ebuilds, since EAPI=4 has unintended side effects
In the creation of the -r1 versions of these files, the EAPI was changed from 2 to 4
for the purpose of using the REQUIRED_USE environment variable.  However, this is now
known to be inappropriate, since it implies other changed expectations, such as the
naming of temporary directories during unpack phase, etc.  This is a quick revert to
EAPI=2 to fix the builders.

BUG=None
TEST=None

Change-Id: I1d9fbfdaaf1de55ba1cff1d2c5461ade6fe5d602
Reviewed-on: https://gerrit.chromium.org/gerrit/19114
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Micah Catlin <micahc@chromium.org>
Reviewed-by: Micah Catlin <micahc@chromium.org>
Tested-by: Micah Catlin <micahc@chromium.org>