Turn on arm hardfp for whole system. Part - 2.
Reintroduced part 3; This is a forced flag day, breaks the tree
for 4 hours.  Landing it since this is aparently known/required (albeit
not fully announced).

Modification includes -
 - added hardfp USE flag to generic-arm and daisy
 - change -mfloat-abi to hard for generic-arm and daisy
 - change -mtune=cortex-a8 to -mtune=cortex-a15 for generic-arm and daisy
 - change -mtune=cortex-a8 to -mtune=cortex-a9 for waluigi

With this CL and "https://gerrit.chromium.org/gerrit/#change,22209",
we can build all packages except media-libs/mali-drivers, which needs
another CL (sheu@) to fix; and gtalk, which will be disabled in
another CL.

BUG=None
TEST=Build the whole system.

Change-Id: I35223c1ec1412bb871990fcbac7fa794446068ca
Reviewed-on: https://gerrit.chromium.org/gerrit/22349
Reviewed-by: Brian Harring <ferringb@chromium.org>
Tested-by: Brian Harring <ferringb@chromium.org>