Hack around bad /lib/firmware permissions for media-libs/mfc-fw
The tarball for media-libs/mfc-fw has bad directory permissions
set which prevent future operations on the directory.  This forces
a fix to those permissions to the portage defaults after the tar
is extracted.

As the upstream tarball includes the -r3 modified I can not
bump the ebuild number so this change won't be automatically
picked up.

BUG=chrome-os-partner:10579
TEST=emerged mfc-fw, verified dir permissions and updat_kernel works

Change-Id: Ia8972122f71f19b9e344f42193053bde2d5da3b8
Reviewed-on: https://gerrit.chromium.org/gerrit/25955
Tested-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Ready: Jon Kliegman <kliegs@chromium.org>