Daisy : add wakeup_input_device_names for keyboard and trackpad
This will tell the power manager to disable the wakeups of the
Cypress trackpad and the chromeos-ec keyboard when the lid
is closed.

BUG=chrome-os-partner:11661
TEST=Boot system.
$ head /var/log/power_manager/powerm.LATEST
Check that the following lines show up:
[0721/205814:INFO:input.cc(237)] Set power/wakeup for input0 state: enabled
[0721/205814:INFO:input.cc(237)] Set power/wakeup for input1 state: enabled
Now, close the lid to suspend, then open to resume.
Check that before suspend the following:
[0721/205936:INFO:input.cc(237)] Set power/wakeup for input1 state: disabled
[0721/205936:INFO:input.cc(237)] Set power/wakeup for input0 state: disabled
And after:
[0721/205945:INFO:input.cc(237)] Set power/wakeup for input1 state: enabled
[0721/205945:INFO:input.cc(237)] Set power/wakeup for input0 state: enabled
Also, close the lid by holding a magnet to the hall effect sensor.
Verify that the system doesn't resume on touchpad and keyboard.

Change-Id: I16eeac641f7be26696e6866bd8954df2c7cbf4c8
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28133
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>