Remove adhd from Beaglebone's package.provided.
This was done as part of an effort to trim the iamge size down,
at a time when this dragged in more dependencies. It's now breaking
the Beaglebone build and the effects of not stubbing it out aren't
as bad as they were previously.

BUG=none
TEST=Build beaglebone image, verify build_packages doesn't fail

Change-Id: I7ae1d5a42ff6b07fbf065c92ff2309faeda8cde7
Reviewed-on: https://gerrit.chromium.org/gerrit/31165
Commit-Ready: Liam McLoughlin <lmcloughlin@chromium.org>
Reviewed-by: Liam McLoughlin <lmcloughlin@chromium.org>
Tested-by: Liam McLoughlin <lmcloughlin@chromium.org>