chromeos-bsp-daisy: Add low_battery_shutdown * configs
For daisy the EC uses a percentage based low battery threshold to determine when
to hard power off the device. This changes tells powerd to not use the default
time threshold and instead use a percentage based one that is higher then the
EC, so that we can power down saftey from software.

BUG=chrome-os-partner:13899
TEST=Confirm that board specific configs are installed.

Change-Id: Ifd755398bc07eeb2b96a73b1a19686625734046a
Reviewed-on: https://gerrit.chromium.org/gerrit/33448
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>