chromeos-bsp-daisy : fix chromeos-ec-i2c name in wakeup_input_device_names
As of this cl : https://gerrit.chromium.org/gerrit/#/c/33571/
The name of the input device for the keyboard has changed from
"chromeos-ec" to "chromeos-ec-i2c"

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:15019
TEST=Boot system.
$ head /var/log/power_manager/powerm.LATEST
Check that the following lines show up:
[0721/205814:INFO:input.cc(237)] Set power/wakeup for input0 state: enabled
[0721/205814:INFO:input.cc(237)] Set power/wakeup for input1 state: enabled
Now, close the lid to suspend, then open to resume.
Check that before suspend the following:
[0721/205936:INFO:input.cc(237)] Set power/wakeup for input1 state: disabled
[0721/205936:INFO:input.cc(237)] Set power/wakeup for input0 state: disabled
And after:
[0721/205945:INFO:input.cc(237)] Set power/wakeup for input1 state: enabled
[0721/205945:INFO:input.cc(237)] Set power/wakeup for input0 state: enabled
Also, close the lid by holding a magnet to the hall effect sensor.
Verify that the system doesn't resume on touchpad and keyboard.

Change-Id: I3934f554685e2d32db3122b4c8df9725d90314b1
Reviewed-on: https://gerrit.chromium.org/gerrit/34911
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>