chromeos-touch-firmware-daisy : Support trial firmware switching
This will pull down a new firmware package,
chromeos-touch-firmware-daisy-2.4-r2.tbz2, which will contain a total
of four touch firmwares. Here is how they are unpacked and symlinked:

/lib/firmware/cyapa.bin ->
	/opt/google/touchpad/firmware/CYTRA-116002-00_2.4.bin
/lib/firmware/cyapa.bin-trial1 ->
	/opt/google/touchpad/firmware/CYTRA-116002-00_2.7.bin
/lib/firmware/cyapa.bin-trial2 ->
	/opt/google/touchpad/firmware/CYTRA-116002-00_2.8.bin
/lib/firmware/cyapa.bin-trial3 ->
	/opt/google/touchpad/firmware/CYTRA-116002-00_2.9.bin

The firmware updater script now also has a --trial flag. When set,
it will use /var/cache/touch_trial/selection to determine which
of the four firmwares above to use. selection may not be an arbitrary
string, as the script will compare the selection to the file names
actually in /lib/firmware above, and will fall back to the default
cyapa.bin if not one of the above.

The firmware updater script has been changed to write the name of the
firmware to the device's update_fw sysfs property.

This CL also add -t option to chromeos-touch-firmwareupdate.conf, meaning
that trial mode is enabled on the firmwareupdate check on every boot.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium-os:37150,chromium-os:36911
TEST=Check that the /var/cache/touch_trial directory and files exist:
1. Boot a system fresh with this change
2. ls /var/cache/touch_trial
3. Check that this directory contains selection and choices
4. cat /var/cache/touch_trial/selection
5. Check that it returns cyapa.bin
6. cat /var/cache/touch_trial/choices
7. Check the following list is printed:
cyapa.bin
cyapa.bin-trial1
cyapa.bin-trial2
cyapa.bin-trial3

Check that changing to a trial firmware works:
1. echo -n "cyapa.bin-trial1" > /var/cache/touch_trial/selection
2. /opt/google/touchpad/firmware/chromeos-touch-firmwareupdate.sh -d cyapa -tr
3. Check that the firmware update succeeds, and that the current firmware is 2.7

Check that writing garbage to the selection falls back to cyapa.bin:
1. echo -n "foobar" > /var/cache/touch_trial/selection
2. /opt/google/touchpad/firmware/chromeos-touch-firmwareupdate.sh -d cyapa -tr
3. Check that this is printed :
"Improper trial firmware name found. Falling back to default."
And also, check that the firmware update succeeds, and we are back at 2.4
4. cat /var/cache/touch_trial/selection
5. Check that selection is now back to cyapa.bin

Change-Id: Ia0e33e328fd0b45c3aa33796c27d6374dfea2a7c
Reviewed-on: https://gerrit.chromium.org/gerrit/39842
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Commit-Ready: Benson Leung <bleung@chromium.org>